UPM Utilities: Added source and utility functions
This commit adds a source file for the upm utilities. This was
important as not all the functions in upm_utilities.h can be inlined.
The implementation of the delay functions now supports zephyr as
the underlying OS as well as linux. Memory allocation functions
have also been introduced as a part of this commit. These
functions abstract the memory allocation functionality for Linux
as well as Zephyr. These functions will be used rather than
OS specific functions.

Signed-off-by: Abhishek Malik <abhishek.malik@intel.com>