urm37: show an example of statically defining mraa resources in a standard way
This should not be merged.  ignore the chnages in urm37.h.  All the
good stuff is in urm37.c.

It is just an example demonstrating how the resources needed by a
driver (FTI) can be statically initialized.

Other options are available - in the future calls (or upmctl) could be
used to query this array and allow changing certain items.

Currently, in the FTI descriptor struct, a list of upm_protocol_t's
are returned.  We could just pass this entire array via this
mechanism.  It would contain the same data, plus all of the specifics
(pin, bus, etc).

Signed-off-by: Jon Trulson <jtrulson@ics.com>