uln200xa: Initial C implementation, example not usable yet
Currently, there are just not enough available GPIOS to properly
driver this from the x86 core.  There is a PR (#25) in ZMRAA that
should fix this problem, but it has not been merged yet.

My understanding is that GPIO is not available from the ARC core at
all yet either, so...  That too may be addressed in PR #25.

Signed-off-by: Jon Trulson <jtrulson@ics.com>