bmp280/bme280 samples: Add #ifdefs to check for correct default init
The bmp/bme280 can operate in SPI or I2C modes.  However, which mode
you can actually use depends on the board type.  So ifdefs have been
added to these samples so that the default initialization is SPI for
the quark core (101), I2C for the ARC core (101_sss) and I2C again for
everything else.

Signed-off-by: Jon Trulson <jtrulson@ics.com>